Moved ReadableStreamReponse model into its own file
